prepare new structure
