+++ msiempy/__version__.py
@@ -1,4 +1,4 @@
 """
 Project version.
 """
-__version__='0.2.5'
+__version__='0.3.0.dev'
