+++ msiempy/core/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+The core objects of the library: session, configuration, abstract query, list and dict.    
+"""
+from .types import NitroList, NitroDict, NitroObject
+from .session import NitroSession, NitroError
+from .query import FilteredQueryList
+from .config import NitroConfig
